requirements.txt
yfinance
pandas
requests
flask
psycopg2-binary
